requirements.txt
weaviate-client==3.26.7
pydantic==2.7.4
python-docx==1.1.2
openpyxl==3.1.5
pandas==2.2.2
numpy==1.26.4
pytesseract==0.3.10
pdfminer.six==20221105
camelot-py[cv]==0.11.0
PyMuPDF==1.24.8
requests==2.32.3
ujson==5.10.0
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
scikit-learn==1.4.2
matplotlib==3.8.4
seaborn==0.13.2
openpyxl-templates==0.2.5
